requirements.txt
Django>=4.2,<5.0
python-telegram-bot>=20.0,<21.0
Pillow>=10.0
psycopg2-binary>=2.9
redis>=5.0
django-celery-beat>=2.5
python-dotenv>=1.0